LP1980978: Improve SIP2 Patron Status Field
Set Y in the "too many items charged" subfield (05) of the SIP2 Patron
Status field if the patron has the PATRON_EXCEEDS_CHECKOUT_COUNT
penalty.

Set Y in the "too many items lost" subfield (09) of the SIP2 Patron
Status field if the patron has the PATRON_EXCEEDS_LOST_COUNT
penalty.

Include the PATRON_EXCEEDS_LONGOVERDUE_COUNT penalty when checking for
the "too many items overdue" subfield (06) of the Patron Status field.

Too test this bug:

1. Find a patron with either too many items checked out or too many
lost items.

2. Use your favorite SIP client to look them up in Evergreen with a
patron information message (63).

3. Verify that the patron comes with their ability to checkout
blocked, i.e there is a Y in field 00 of the Patron Status field of
the message 64 returned by the SIP server.

4. Verify that fields 05 and 09 are both blank even though one of them
should have a Y in it.

After applying the patch, begin with step 2, above.  When you get to step
4, either 05 or 09 (possilby both) should have a Y in it depending on
the patron's penalties.

Repeat the above steps with a patron who has the
PATRON_EXCEEDS_LONGOVERDUE_COUNT penalty but not the
PATRON_EXCEEDS_OVERDUE_COUNT penalty.  You will look for a Y or a
space in subfield 06.

Signed-off-by: Jason Stephenson <jason@sigio.com>