Add 1355 to custom upgrade
I rebased the egdev branch on rel_3_10 to cleanup some commits that
were duplicated when I tried a merge with what was going to be the
production branch to pick some additional commits from that branch.

The rebase introduced a new DB upgrade script, so this commit adds it
to the custom db upgrade.