Start basic introductory material
Add Start.md as the main document with links to the various concepts
and examples.

Add "What Is Git?" concept document.

Add stubs for the "Getting Git" and "Getting Help" concept documents.

Add link to the interactive rebase example to the Examples section of
the start document.

Link the start document from the README.

Signed-off-by: Jason Stephenson <jason@sigio.com>