Update Alias examples
Replace a copied alias with a shorter one of my own that I actually
added to my own gitconfig earlier today because I found it handier to
get a quick overview of the last few commits to a branch.

Add an example of what that alias looks like in the .gitconfig file.

Add some better examples for git commit.

Add an example of removing an alias via the git config command line.

Signed-off-by: Jason Stephenson <jason@sigio.com>